Quick fix for Mini Racer problem
WARNING: V8 isolate was forked, it can not be disposed and memory will not be reclaimed till the Ruby process exits.
It is VERY likely your process will hang.
If you wish to use v8 in forked environment please ensure the platform is initialized with:
MiniRacer::Platform.set_flags! :single_threaded